Feat : Member 기본 셋팅
