+++ src/main/java/com/prgrms/ijuju/domain/example/dto/request/SampleRequestDTO.java
@@ -1,0 +1,15 @@
+package com.prgrms.ijuju.domain.example.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class SampleRequestDTO {
+    private Long ssn;
+    private String name;
+}
