+++ src/main/java/com/prgrms/ijuju/domain/example/exception/SampleException.java
@@ -1,0 +1,5 @@
+package com.prgrms.ijuju.domain.example.exception;
+
+public enum SampleException {
+
+}
