+++ src/main/java/com/prgrms/ijuju/domain/example/repository/SampleRepository.java
@@ -1,0 +1,7 @@
+package com.prgrms.ijuju.domain.example.repository;
+
+import com.prgrms.ijuju.domain.example.entity.Sample;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SampleRepository extends JpaRepository<Sample, String> {
+}
