+++ src/main/java/com/prgrms/ijuju/domain/example/service/SampleService.java
@@ -1,0 +1,10 @@
+package com.prgrms.ijuju.domain.example.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class SampleService {
+    public void sampleMethod(){
+        System.out.println("SampleService's sampleMethod()");
+    }
+}
