+++ src/main/java/com/prgrms/ijuju/domain/minigame/flipcard/dto/response/FlipCardResponseDTO.java
@@ -1,0 +1,12 @@
+package com.prgrms.ijuju.domain.minigame.flipcard.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+public class FlipCardResponseDTO {
+}
