+++ src/main/java/com/prgrms/ijuju/domain/minigame/oxquiz/dto/request/OXQuizRequestDTO.java
@@ -1,0 +1,12 @@
+package com.prgrms.ijuju.domain.minigame.oxquiz.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+public class OXQuizRequestDTO {
+}
