+++ src/main/java/com/prgrms/ijuju/domain/minigame/oxquiz/entity/Member.java
@@ -1,0 +1,36 @@
+package com.prgrms.ijuju.domain.minigame.oxquiz.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.*;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@ToString
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Table(name="member")
+@EntityListeners(value = { AuditingEntityListener.class })
+public class Member {
+    @Id
+    private String mid;
+    private String mpw;
+    private String mname;
+    private String email;
+    private String role;
+
+    @CreatedDate
+    private LocalDateTime joinDate;
+
+    @LastModifiedDate
+    private LocalDateTime modifiedDate;
+
+}
