+++ src/main/java/com/prgrms/ijuju/domain/stock/mid/dto/response/MidResponseDTO.java
@@ -1,0 +1,12 @@
+package com.prgrms.ijuju.domain.stock.mid.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+public class MidResponseDTO {
+}
