--- src/main/java/com/prgrms/ijuju/example/dto/request/SampleRequestDTO.java
@@ -1,15 +1,0 @@
-package com.prgrms.ijuju.example.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class SampleRequestDTO {
-    private Long ssn;
-    private String name;
-}
