--- src/main/java/com/prgrms/ijuju/example/exception/SampleException.java
@@ -1,5 +1,0 @@
-package com.prgrms.ijuju.example.exception;
-
-public enum SampleException {
-
-}
