--- src/main/java/com/prgrms/ijuju/example/repository/SampleRepository.java
@@ -1,7 +1,0 @@
-package com.prgrms.ijuju.example.repository;
-
-import com.prgrms.ijuju.example.entity.Sample;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SampleRepository extends JpaRepository<Sample, String> {
-}
