--- src/main/java/com/prgrms/ijuju/example/service/SampleService.java
@@ -1,10 +1,0 @@
-package com.prgrms.ijuju.example.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class SampleService {
-    public void sampleMethod(){
-        System.out.println("SampleService's sampleMethod()");
-    }
-}
