--- src/main/java/com/prgrms/ijuju/minigame/flipcard/dto/response/FlipCardResponseDTO.java
@@ -1,12 +1,0 @@
-package com.prgrms.ijuju.minigame.flipcard.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-public class FlipCardResponseDTO {
-}
