--- src/main/java/com/prgrms/ijuju/minigame/lottery/dto/response/LotteryResponseDTO.java
@@ -1,12 +1,0 @@
-package com.prgrms.ijuju.minigame.lottery.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-public class LotteryResponseDTO {
-}
