--- src/main/java/com/prgrms/ijuju/stock/advanced/dto/response/AdvancedResponseDTO.java
@@ -1,12 +1,0 @@
-package com.prgrms.ijuju.stock.advanced.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-public class AdvancedResponseDTO {
-}
