--- src/main/java/com/prgrms/ijuju/stock/begin/dto/request/BeginRequestDTO.java
@@ -1,12 +1,0 @@
-package com.prgrms.ijuju.stock.begin.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-public class BeginRequestDTO {
-}
