--- src/main/java/com/prgrms/ijuju/stock/begin/dto/response/BeginResponseDTO.java
@@ -1,12 +1,0 @@
-package com.prgrms.ijuju.stock.begin.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-public class BeginResponseDTO {
-}
